+++ aqua/diagnostics/timeseries/base.py
@@ -7,7 +7,7 @@ import xarray as xr
 from aqua.core.fixer import EvaluateFormula
 from aqua.core.logger import log_configure
 from aqua.core.util import frequency_string_to_pandas, pandas_freq_to_string, strlist_to_phrase, time_to_string
-from aqua.diagnostics.base import SAVE_FORMAT, Diagnostic, OutputSaver, TitleBuilder, start_end_dates
+from aqua.diagnostics.base import SAVE_FORMAT, Diagnostic, OutputSaver, TitleBuilder
 
 xr.set_options(keep_attrs=True)
 
@@ -54,25 +54,22 @@ class BaseMixin(Diagnostic):
             lat_limits (list): The latitude limits to be used. Overriden by region.
             loglevel (str): The log level to be used. Default is 'WARNING'.
         """
-        super().__init__(catalog=catalog, model=model, exp=exp, source=source, regrid=regrid, loglevel=loglevel)
+        super().__init__(
+            catalog=catalog,
+            model=model,
+            exp=exp,
+            source=source,
+            regrid=regrid,
+            startdate=startdate,
+            enddate=enddate,
+            std_startdate=std_startdate,
+            std_enddate=std_enddate,
+            loglevel=loglevel,
+        )
 
         # Log name is the diagnostic name with the first letter capitalized
         self.logger = log_configure(log_level=loglevel, log_name=diagnostic_name.capitalize())
         self.diagnostic_name = diagnostic_name
-
-        # We want to make sure we retrieve the required amount of data with a single Reader instance
-        self.startdate, self.enddate = start_end_dates(
-            startdate=startdate, enddate=enddate, start_std=std_startdate, end_std=std_enddate
-        )
-        # They need to be stored to evaluate the std on the correct period
-        self.std_startdate = self.startdate if std_startdate is None else std_startdate
-        self.std_enddate = self.enddate if std_enddate is None else std_enddate
-        # Finally we need to set the start and end dates of the data
-        self.plt_startdate = startdate
-        self.plt_enddate = enddate
-        self.logger.debug(f"Retrieve start date: {self.startdate}, End date: {self.enddate}")
-        self.logger.debug(f"Plot start date: {self.plt_startdate}, End date: {self.plt_enddate}")
-        self.logger.debug(f"Std start date: {self.std_startdate}, Std end date: {self.std_enddate}")
 
         # Set the region based on the region name or the lon and lat limits
         self.region, self.lon_limits, self.lat_limits = self._set_region(
@@ -114,45 +111,89 @@ class BaseMixin(Diagnostic):
         # of all the variables
         if formula:
             super().retrieve(reader_kwargs=reader_kwargs, months_required=self.MINIMUM_MONTHS_REQUIRED)
+            source_data = self.data
+            source_std_data = self.std_data
             self.logger.debug("Evaluating formula %s", var)
             self.data = EvaluateFormula(
-                data=self.data, formula=var, long_name=long_name, short_name=short_name, units=units, loglevel=self.loglevel
+                data=source_data,
+                formula=var,
+                long_name=long_name,
+                short_name=short_name,
+                units=units,
+                loglevel=self.loglevel,
             ).evaluate()
             if self.data is None:
                 raise ValueError(f"Error evaluating formula {var}. Check the variable names and the formula syntax.")
+            self._copy_aqua_attrs(source=source_data, target=self.data)
+            if self.std_data is not None:
+                self.std_data = EvaluateFormula(
+                    data=source_std_data,
+                    formula=var,
+                    long_name=long_name,
+                    short_name=short_name,
+                    units=units,
+                    loglevel=self.loglevel,
+                ).evaluate()
+                if self.std_data is None:
+                    raise ValueError(f"Error evaluating formula {var}. Check the variable names and the formula syntax.")
+                self._copy_aqua_attrs(source=source_std_data, target=self.std_data)
         else:
             super().retrieve(var=var, reader_kwargs=reader_kwargs, months_required=self.MINIMUM_MONTHS_REQUIRED)
             if self.data is None:
                 raise ValueError(f"Variable {var} not found in the data. Check the variable name and the data source.")
             # Get the xr.DataArray to be aligned with the formula code
-            self.data = self.data[var]
-
-        if self.plt_startdate is None:
-            self.plt_startdate = self.data.time.min().values
-            self.logger.debug("Plot start date set to %s", self.plt_startdate)
-        if self.plt_enddate is None:
-            self.plt_enddate = self.data.time.max().values
-            self.logger.debug("Plot end date set to %s", self.plt_enddate)
+            source_data = self.data
+            self.data = source_data[var]
+            self._copy_aqua_attrs(source=source_data, target=self.data)
+            if self.std_data is not None:
+                source_std_data = self.std_data
+                self.std_data = source_std_data[var]
+                self._copy_aqua_attrs(source=source_std_data, target=self.std_data)
 
         # Customization of the data, expecially needed for formula
         if units is not None:
-            self._check_data(var, units)
+            self.data = super()._check_data(data=self.data, var=var, units=units)
+            if self.std_data is not None:
+                self.std_data = super()._check_data(data=self.std_data, var=var, units=units)
         if long_name is not None:
             self.data.attrs["long_name"] = long_name
+            if self.std_data is not None:
+                self.std_data.attrs["long_name"] = long_name
         # We want to be sure that a long_name is always defined for description setup
         elif self.data.attrs.get("long_name") is None:
             self.data.attrs["long_name"] = var
+            if self.std_data is not None and self.std_data.attrs.get("long_name") is None:
+                self.std_data.attrs["long_name"] = var
         # We use the short_name as the name of the variable
         # to be always used in plots
         if short_name is not None:
             self.data.attrs["short_name"] = short_name
             self.data.name = short_name
+            if self.std_data is not None:
+                self.std_data.attrs["short_name"] = short_name
+                self.std_data.name = short_name
         else:
             self.data.attrs["short_name"] = var
+            if self.std_data is not None:
+                self.std_data.attrs["short_name"] = var
+
+    @staticmethod
+    def _copy_aqua_attrs(source, target):
+        """Copy AQUA_* attributes from source to target without overriding existing values."""
+        if source is None or target is None:
+            return
+
+        source_attrs = getattr(source, "attrs", {})
+        target_attrs = getattr(target, "attrs", {})
+        for key, value in source_attrs.items():
+            if key.startswith("AQUA_") and key not in target_attrs:
+                target.attrs[key] = value
 
     def compute_std(self, freq: str, exclude_incomplete: bool = True, center_time: bool = True, box_brd: bool = True):
         """
         Compute the standard deviation of the data. Support for monthly and annual frequencies.
+        The standard deviation is computed over the time window defined by std_startdate and std_enddate attributes.
+        If these attributes are not defined, we raise an error, since we need a time window to compute the standard deviation.
 
         Args:
             freq (str): The frequency to be used for the resampling.
@@ -165,6 +206,9 @@ class BaseMixin(Diagnostic):
             self.logger.error("Frequency not provided")
             raise ValueError("Frequency not provided")
 
+        if self.std_startdate is None or self.std_enddate is None:
+            raise ValueError("Standard deviation start and end dates must be provided to compute the standard deviation.")
+
         freq = frequency_string_to_pandas(freq)
         str_freq = pandas_freq_to_string(freq)
         self.logger.info("Computing %s standard deviation", str_freq)
@@ -176,7 +220,7 @@ class BaseMixin(Diagnostic):
             "annual": {"data": self.annual, "groupdby": None},
         }
 
-        data = self.data
+        data = self.std_data
         data = self.reader.fldmean(data, box_brd=box_brd, lon_limits=self.lon_limits, lat_limits=self.lat_limits)
         data = self.reader.timmean(data, freq=freq, exclude_incomplete=exclude_incomplete, center_time=center_time)
 
@@ -185,11 +229,6 @@ class BaseMixin(Diagnostic):
             self.logger.warning(f"Not enough data to compute {str_freq} standard deviation")
             data = None
         else:
-            data = data.sel(time=slice(self.std_startdate, self.std_enddate))
-            if self.std_startdate is None or self.std_enddate is None:
-                self.std_startdate = data.time.min().values
-                self.std_enddate = data.time.max().values
-            self.logger.debug(f"Standard deviation period from {self.std_startdate} to {self.std_enddate}")
             if freq_dict[str_freq]["groupdby"] is not None:
                 data = data.groupby(freq_dict[str_freq]["groupdby"]).std("time")
             else:  # For annual data, we compute the std over all years
@@ -197,11 +236,6 @@ class BaseMixin(Diagnostic):
 
             if self.region is not None:
                 data.attrs["AQUA_region"] = self.region
-
-            # Store start and end dates for the standard deviation.
-            # pd.Timestamp cannot be used as attribute, so we convert to a string
-            data.attrs["std_startdate"] = time_to_string(self.std_startdate)
-            data.attrs["std_enddate"] = time_to_string(self.std_enddate)
 
             # Load data in memory for faster plot
             self.logger.debug(f"Loading std data for frequency {str_freq} in memory")
@@ -292,16 +326,6 @@ class BaseMixin(Diagnostic):
                 rebuild=rebuild,
                 extra_keys=extra_keys,
             )
-
-    def _check_data(self, var: str, units: str):
-        """
-        Make sure that the data is in the correct units.
-
-        Args:
-            var (str): The variable to be checked.
-            units (str): The units to be checked.
-        """
-        self.data = super()._check_data(data=self.data, var=var, units=units)
 
 
 class PlotBaseMixin: