+++ aqua/diagnostics/timeseries/cli_timeseries.py
@@ -51,7 +51,6 @@ def main(argv=None):
             diagnostic_name = cli.config_dict["diagnostics"]["timeseries"].get("diagnostic_name", "timeseries")
             center_time = cli.config_dict["diagnostics"]["timeseries"].get("center_time", True)
             exclude_incomplete = cli.config_dict["diagnostics"]["timeseries"].get("exclude_incomplete", True)
-            extend = cli.config_dict["diagnostics"]["timeseries"].get("extend", True)
 
             for var in cli.config_dict["diagnostics"]["timeseries"].get("variables", []):
                 var_config, regions = load_var_config(
@@ -81,7 +80,6 @@ def main(argv=None):
                             "rebuild": cli.rebuild,
                             "center_time": center_time,
                             "exclude_incomplete": exclude_incomplete,
-                            "extend": extend,
                         }
 
                         # Initialize a list of len from the number of datasets
@@ -98,9 +96,9 @@ def main(argv=None):
                             )
 
                         # Reference datasets are evaluated on the maximum time range of the datasets
-                        startdate = round_startdate(pd.Timestamp(min(t.plt_startdate for t in ts)))
-                        enddate = round_enddate(pd.Timestamp(max(t.plt_enddate for t in ts)))
-                        cli.logger.info(f"Start date: {startdate}, End date: {enddate}")
+                        startdate = round_startdate(pd.Timestamp(min(t.startdate for t in ts)))
+                        enddate = round_enddate(pd.Timestamp(max(t.enddate for t in ts)))
+                        cli.logger.info(f"Total start date: {startdate}, end date: {enddate}")
 
                         # Initialize a list of len from the number of references
                         if "references" in cli.config_dict:
@@ -177,7 +175,6 @@ def main(argv=None):
                 diagnostic_name = cli.config_dict["diagnostics"]["timeseries"].get("diagnostic_name", "timeseries")
                 center_time = cli.config_dict["diagnostics"]["timeseries"].get("center_time", True)
                 exclude_incomplete = cli.config_dict["diagnostics"]["timeseries"].get("exclude_incomplete", True)
-                extend = cli.config_dict["diagnostics"]["timeseries"].get("extend", True)
 
                 for region in regions:
                     try:
@@ -195,7 +192,6 @@ def main(argv=None):
                             "rebuild": cli.rebuild,
                             "center_time": center_time,
                             "exclude_incomplete": exclude_incomplete,
-                            "extend": extend,
                         }
 
                         # Initialize a list of len from the number of datasets
@@ -211,8 +207,8 @@ def main(argv=None):
                             )
 
                         # Reference datasets are evaluated on the maximum time range of the datasets
-                        startdate = pd.Timestamp(min(t.plt_startdate for t in ts))
-                        enddate = pd.Timestamp(max(t.plt_enddate for t in ts))
+                        startdate = pd.Timestamp(min(t.startdate for t in ts))
+                        enddate = pd.Timestamp(max(t.enddate for t in ts))
 
                         # Initialize a list of len from the number of references
                         if "references" in cli.config_dict:
@@ -324,8 +320,8 @@ def main(argv=None):
                             )
 
                         # Reference datasets are evaluated on the maximum time range of the datasets
-                        startdate = pd.Timestamp(min(t.plt_startdate for t in ts))
-                        enddate = pd.Timestamp(max(t.plt_enddate for t in ts))
+                        startdate = pd.Timestamp(min(t.startdate for t in ts))
+                        enddate = pd.Timestamp(max(t.enddate for t in ts))
 
                         # Initialize a list of len from the number of references
                         if "references" in cli.config_dict: